Add merge in button to the table.
This fits into this PR because it is part of refactoring the datacard
toolbar.
